Move onUnitsChange to PlotterBase to remove code duplication
